requirements.txt
Flask==3.1.0
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
pypdf==5.4.0
python-docx==1.1.2
requests==2.32.3
scikit-learn==1.6.1
